moved context manager into separate file
